A safer way of depickling petastorm objects from legacy package names (#26)
* A safer way of depickling petastorm objects from legacy package names

We have existing serialized instances of datasets in existing files. Prevoius way of
replacing sys.module[] dictionary entries had a side effect: it was not possible to
pass a unischema deserailized in this fashion to a pyspark executor.

The new way simply replaces module names in the serialized stream in the safest way I think is possible.