Fix channel ordering of encoded images.
We expect all inputs/outputs to the codec to be in RGB format. When we introduced opencv codec, we did not account for the fact that opencv expects all
its input/output images to be in BGR. This resulted in the serialized png streams become a non standard BGR streams. Also, it broke compatibility with older datasets that
wrote-out standard RGB PNG.

Fixing this issue by converting RGB impages into BGR before encoding. Reversing the BGR to RGB when decoding.