[Part 2] New reader: ShufflingBuffer implementations (Random and Noop)
Items can be added to a shuffling buffer and removed in a different order as defined by the concrete shuffling algorithm.
Currently we have a `NoopShufflingBuffer` implementation which is not really shuffling, but just a FIFO implementation. Used for tests and when we don't want to shuffle.
`RandomShufflingBuffer` implements a random shuffling buffer with low and high water marks.