Resurrect Decimal type and make pyarrow.serialize optional
Replacing Decimal with strings at the codec level was a bad idea. It broke a bunch of existing code.
Among others, it also broke ngrams which were relying on Decimal field for threashold checks.

In this commit, we add pyarrow_serliaze optional argument to the ProcessPool constructor, so a user
that needs pyarrow fast serialization, but is willing to sacrifice some types that can be used
in a Unischema can get faster serialization.