Add '_max_workers` property to `SameThreadExecutor`
For now we define _max_workers and polymorphically with `ThreadPoolExecutor` and `ProcessPoolExecutor` in
`worker_loop`. We should not rely on private `_max_workers` for that, but first we need to refactor
`ReaderV2` so we would know during construction of the reader what is the number of workers. Currently we
just get an `*PoolExecutor` instance which don't have a public API for that.