Support _common_metadata generation for hive parquet stores
An example of a Hive parquet store that I saw, had  `metadata` field with `None` value  in the
`pyarrow.parquet.read_metadata(f).schema.to_arrow_schema()`. In this commit, we remove an assumption
on any kind of a dictionary in that property and create a dictionary ourselves, if none could be loaded.