Make test_shuffle_drop_ratio by not setting shuffle_row_drop_partitions >> rowgroup size.
It was working fast while we knew the number of rows in each rowgroup. We don't have this information now,
so there are many row-groups that are dropped without being used by shuffle_row_drop_partition feature.

The time it takes for the test to finish now is 2 seconds instead of 23 seconds (and I assume it takes more on travis)