Do not attempt to override pyarrow's dataset.partitions attribute
Removing a hack where we manually override pyarrow's dataset.partitions
property because of an assumption of other parts of code that the
attribute is always set and iteratable.

pyarrow 5.0 breaks as it protects the attribute against modification.

We fix the issue by explicitly checking for dataset.partitions being
None in relevant parts of petastorm code.

Add a unittest configuration to cover pyarrow 3, 4, and 5.